05/09/2025

+++ adminPanel/controller/adminCtrl.js
@@ -3,39 +3,55 @@ const moment = require('moment');
 const { all } = require('../routes/adminRoutes');
 const fs = require('fs')
 const path = require('path')
+const bcrypt = require('bcrypt')
 
 module.exports.admin = async (req, res) => {
   try {
-    return res.render('dashboard')
+    if (req.cookies.admin && req.cookies.admin._id) {
+      return res.render('dashboard')
+    }
+    else {
+      return res.redirect('/')
+    }
   } catch (err) {
     console.log(err);
-    return res.redirect('/admin')
+    return res.redirect('/')
 
   }
 }
 
 module.exports.view = async (req, res) => {
   try {
-        var search ="";
-       if(req.query.search) {
+
+    if (req.cookies.admin && req.cookies.admin._id) {
+      var search = "";
+      if (req.query.search) {
         search = req.query.search
       }
       let data = await adminModel.find({
-         $or : [
-        {
-          name : {$regex : search,$options : 'i' }
-        },
-        {
-          email : {$regex : search,$options : 'i' }
-        },
-        {
-          city : {$regex : search,$options : 'i' }
-        }
-      ]  
+        $or: [
+          {
+            name: { $regex: search, $options: 'i' }
+          },
+          {
+            email: { $regex: search, $options: 'i' }
+          },
+          {
+            city: { $regex: search, $options: 'i' }
+          }
+        ]
       })
-    return res.render('view', {
-      data
-    })
+      return res.render('view', {
+        data
+      })
+    }
+    else {
+      return res.redirect('/')
+    }
+
+
+
+
   }
   catch (err) {
     console.log(err);
@@ -62,7 +78,12 @@ module.exports.viewSingle = async (req, res) => {
 
 module.exports.addAdmin = async (req, res) => {
   try {
-    return res.render('addAdmin')
+    if (req.cookies.admin && req.cookies.admin._id) {
+      return res.render('addAdmin')
+    }
+    else {
+      return res.redirect('/')
+    }
   } catch (err) {
     console.log(err);
     return res.redirect('/admin')
@@ -72,16 +93,23 @@ module.exports.addAdmin = async (req, res) => {
 
 module.exports.addForm = async (req, res) => {
   try {
-    req.body.name = req.body.fname + " " + req.body.lname
-    req.body.profile = req.file.filename
-    req.body.status = true
-    req.body.start_date = moment().format('MM Do YYYY, h:mm:ss a');
-    req.body.update_date = moment().format('MM Do YYYY, h:mm:ss a');
+    if (req.cookies.admin && req.cookies.admin._id) {
+      req.body.name = req.body.fname + " " + req.body.lname
+      req.body.profile = req.file.filename
+      req.body.status = true
+      req.body.start_date = moment().format('MM Do YYYY, h:mm:ss a');
+      req.body.update_date = moment().format('MM Do YYYY, h:mm:ss a');
+      req.body.password = bcrypt.hashSync(req.body.password, 10);
 
-    // console.log(req.body);
+      // console.log(req.body);
 
-    await adminModel.create(req.body)
-    return res.redirect('/admin/view')
+      await adminModel.create(req.body)
+      return res.redirect('/admin/view')
+
+    }
+    else {
+      return res.redirect('/')
+    }
   } catch (err) {
     console.log(err);
     return res.redirect('/admin')
@@ -92,6 +120,10 @@ module.exports.addForm = async (req, res) => {
 
 module.exports.delete = async (req, res) => {
   try {
+
+    if (req.cookies.admin && req.cookies.admin._id) {
+
+    }
     // find the record which want to delete
     const single = await adminModel.findById(req.params.id)
 
@@ -104,6 +136,11 @@ module.exports.delete = async (req, res) => {
       await adminModel.findByIdAndDelete(req.params.id)
       return res.redirect('/admin/view')
     }
+    else {
+      return res.redirect('/')
+    }
+
+
   } catch (err) {
     console.log(err);
     return res.redirect('/admin/view')
@@ -114,16 +151,24 @@ module.exports.delete = async (req, res) => {
 module.exports.editAdmin = async (req, res) => {
 
   try {
-    const single = await adminModel.findById(req.params.id)
-    if (single) {
-      return res.render('editAdmin', {
-        single
-      })
+
+    if (req.cookies.admin && req.cookies.admin._id) {
+
+      const single = await adminModel.findById(req.params.id)
+      if (single) {
+        return res.render('editAdmin', {
+          single
+        })
+      }
+      else {
+        console.log('No Record Found !!');
+        return res.redirect('/view')
+      }
     }
     else {
-      console.log('No Record Found !!');
-      return res.redirect('/view')
+      return res.redirect('/')
     }
+
   } catch (err) {
     console.log(err);
     return res.redirect('/admin')
@@ -133,25 +178,33 @@ module.exports.editAdmin = async (req, res) => {
 
 module.exports.edit = async (req, res) => {
   try {
-    const single = await adminModel.findById(req.params.id)
-    console.log(single.profile);
-    if (req.body) {
-      if (req.file) {
-        if (single.profile) {
-          let oldpath = path.join(__dirname, '../uploads/', single.profile)
-          fs.unlinkSync(oldpath)
+
+    if (req.cookies.admin && req.cookies.admin._id) {
+
+      const single = await adminModel.findById(req.params.id)
+      console.log(single.profile);
+      if (req.body) {
+        if (req.file) {
+          if (single.profile) {
+            let oldpath = path.join(__dirname, '../uploads/', single.profile)
+            fs.unlinkSync(oldpath)
+          }
+          req.body.profile = req.file.filename;
+          req.body.name = req.body.fname + " " + req.body.lname;
+          req.body.update_date = moment().format('MM Do YYYY, h:mm:ss a');
         }
-        req.body.profile = req.file.filename;
-        req.body.name = req.body.fname + " " + req.body.lname;
-        req.body.update_date = moment().format('MM Do YYYY, h:mm:ss a');
+        await adminModel.findByIdAndUpdate(req.params.id, req.body)
+
+      } else {
+        console.log("No Data Found !!");
+
       }
-      await adminModel.findByIdAndUpdate(req.params.id, req.body)
-
-    } else {
-      console.log("No Data Found !!");
-
+      return res.redirect('/admin/view')
     }
-    return res.redirect('/admin/view')
+    else {
+      return res.redirect('/')
+    }
+
   } catch (err) {
     console.log(err);
     return res.redirect('/admin')
